test_script created
I got the code in the make_enhanced.py file to write the required code in enhanced.py to create a test_script.py. However, it is literally just rewriting enhanced.py to a new file and renaming it. There is currently nothing being changed that I can tell, and that is an issue and things should be changing.